+++ packages/styleguide/src/examples/examples.module.ts
@@ -4,7 +4,6 @@ import { AuiModule } from '../aui/aui.module';
 import { RouterModule } from '@angular/router';
 import { EXAMPLES_ROUTES } from './examples.routes';
 import { CodeSnippetDemoPageComponent } from './pages/aui-code-snippet/aui-snippet.page';
-import { AgendaDemoPageComponent } from './pages/aui-agenda/aui-agenda.page';
 import { AvatarDemoPageComponent } from './pages/aui-avatar/aui-avatar.page';
 import { CalendarDemoPage } from './pages/aui-calendar/aui-calendar.page';
 import { ContextDemoPageComponent } from './pages/aui-context/aui-context';
@@ -43,7 +42,6 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
     declarations: [
         LogoDemoPageComponent,
         CodeSnippetDemoPageComponent,
-        AgendaDemoPageComponent,
         AvatarDemoPageComponent,
         CalendarDemoPage,
         ContextDemoPageComponent,