+++ packages/ngx-leaflet/src/lib/services/map.service.ts
@@ -1,6 +1,8 @@
-declare var require: any;
 import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { isPlatformBrowser } from '@angular/common';
+import * as L from 'leaflet';
+import * as eL from 'esri-leaflet';
+import 'leaflet-draw';
 
 @Injectable()
 export class MapService {
@@ -14,9 +16,8 @@ export class MapService {
 
   requireMap() {
     if (isPlatformBrowser(this.platformId)) {
-      this.L = require('leaflet');
-      this.esri = require('esri-leaflet');
-      require('leaflet-draw');
+      this.L = L;
+      this.esri = eL;
     }
   }